[Refactoring] Extract CollapseNestedIfStmt to its own file
